refactor(plugins): 新增 plugin-handlers-utils 去重 handler，依赖卫生
- 新建 @covel/plugin-handlers-utils，去重 5 个 handler 共 ~193 行重复
- 依赖卫生：narrator 删冗余 devDep，npc-graph/world-init 补 zod 声明
- runner 638→364、create-world 602→276 拆分
- 同步 CLAUDE.md 包数 14→15
- create 13 + test-runtime 20 + 各插件测试全过
