+++ packages/plugin-handlers-utils/package.json
@@ -1,0 +1,25 @@
+{
+  "name": "@covel/plugin-handlers-utils",
+  "version": "0.0.4",
+  "private": true,
+  "type": "module",
+  "exports": {
+    ".": {
+      "types": "./src/index.ts",
+      "import": "./src/index.ts"
+    }
+  },
+  "scripts": {
+    "test": "vitest run --passWithNoTests",
+    "test:watch": "vitest",
+    "lint": "tsc --noEmit",
+    "clean": "node ../../scripts/remove-paths.mjs dist"
+  },
+  "dependencies": {
+    "@covel/shared": "workspace:*"
+  },
+  "devDependencies": {
+    "typescript": "^6.0.3",
+    "vitest": "^4.1.4"
+  }
+}
