+++ plugins/character-presence/package.json
@@ -8,6 +8,7 @@
     "test:watch": "vitest"
   },
   "dependencies": {
+    "@covel/plugin-handlers-utils": "workspace:*",
     "@covel/tools": "workspace:*"
   },
   "devDependencies": {
