+++ plugins/narrator/package.json
@@ -8,9 +8,6 @@
     "test:watch": "vitest"
   },
   "devDependencies": {
-    "@covel/context": "workspace:*",
-    "@covel/plugin-loader": "workspace:*",
-    "@covel/shared": "workspace:*",
     "vitest": "^4.1.4"
   }
 }
