+++ plugins/npc-graph/package.json
@@ -7,6 +7,9 @@
     "test": "vitest run",
     "test:watch": "vitest"
   },
+  "dependencies": {
+    "zod": "^4.3.6"
+  },
   "devDependencies": {
     "@covel/plugin-loader": "workspace:*",
     "@covel/plugin-test-utils": "workspace:*",
