+++ plugins/world-init/package.json
@@ -8,7 +8,8 @@
     "test:watch": "vitest"
   },
   "dependencies": {
-    "@covel/tools": "workspace:*"
+    "@covel/tools": "workspace:*",
+    "zod": "^4.3.6"
   },
   "devDependencies": {
     "vitest": "^4.1.4"
